+++ internal/console/hostinfo_stub.go
@@ -5,3 +5,7 @@ package console
 func snapshotHost(_ *cpuSampler, _ string) SystemView {
 	return baseSystemIdentity()
 }
+
+func snapshotHostPlaque(sampler *cpuSampler, colonyRoot string) SystemView {
+	return snapshotHost(sampler, colonyRoot)
+}